Added routing.
Had to add a new component, RealHome. This is to store the 3 components of the home page. We need this because before the app component was holding these 3 components, but now the app component holds the routing. The two routes are / and /details/stockSymbol, which routes to detailedPage. The stockSymbol is a parameter that is given to detailedPage, and is stored as stockSymbol within that component using the getparams function.