completed the menu
the menu is done, changed all the name in the header to appropriate named . plus change the header to scroll down when scrolling down. finally made a small change to Utility.js to take how many stocks to  pull from the database in no input it will defult to 6 stocks as it was.
 needed to double checked from Dillon to make sure it  gets the data from the database since i can not connect to the data base, but ii think it should work without any chages