Mocked API kindof
Made a module to mock the api. Replaced all references of 'axios' with references to apiObject, which is a created instance of axios, so hopefully that would make mocking work everywhere, because apiObject is mocked, and since apiObject is used everywhere, that is how it should work.
Most of the tests fail. This is because the api isn't being mocked properly, IDK why that is occurring.